requirements.txt
fastapi
uvicorn
requests
requests-toolbelt
pydantic
motor
pyPDF2
python-dotenv
langchain-core
langchain-groq
langchain-huggingface
langchain-community
sentence-transformers
transformers
speechrecognition
pillow
pytest
pytest-asyncio
python-multipart
torch
torchaudio
soundfile
librosa
pydub
speechrecognition
pymongo
nltk
spacy
tqdm
jinja2
python-jose[cryptography]
passlib[bcrypt]
passlib